Simplify explicit-state update client internals
